chore: extract brokerProbe/webullToken/events from dashboard index (move-only)
